Adds memcached to the setup
Sessions are better handled by CKAN by memcached rather than storing
files on disk (in the /tmp directory).  This installs memcached and sets
the config to point to it.

For this to work, it also needs the version of ckanext-nhsengland that
contains pylibmc as a dependency, so you should try them out together.