Add project level ansible config
This allows us to enable pipelining fixing an error introduced with the
move to become[_user].  See [the docs][1] for a long form explanation on
why this was necessary.  We hit this with the upgrade to Ansible 2.4 in
the db playbook, specifically when trying to create a user in Postgres.

[1]: http://docs.ansible.com/ansible/latest/become.html#becoming-an-unprivileged-user