Bootstrap python and install a modern pip
pip on 12.04 is too old to connect to current day PyPI since HTTPS
connections are enforced.  Upgrading to a modern (and non-OS managed)
pip only fixes part of the problem.  12.04 ships with an OpenSSL that's
now too out of date to connect to PyPI.  Rather than hacking around that
only to perform the upgrade to 16.04 later I'm starting that upgrade
here and now.  However 16.04 doesn't ship with a 2.x python so we have
to bootstrap that first.