Install Jetty and Solr before Apache
Somewhere in the intersection of newer Ubuntu and Jetty/Java either the
default port for Jetty changed or the deb started Jetty automatically.
Jetty8's default port is 8080, the same as we want Apache to be.  Since
we don't want Jetty running on that port anyway we can install and
configure it before Apache to avoid the problem.  The only knock-on
effect is having to copy the Solr template to our Ckan install directory
after Ckan has been installed.